requirements.txt
flask
websocket-client
requests
gunicorn
gevent